Fixed physics when scaling the camera (finally)! Cleaned up and made some more mess.
The physics was being wrecked by the .world property on sprite bodies. It was calculated by summing the camera coordinate with the world's transform translation (ouch). Removing this, and no longer using the _fixX/_fixY methods in TilemapLayer fixes collision completely - between objects and between tilemaps - when scaling the camera. Great success.